Add kernel port of dhcpcd.  It's now possible to configure TCP/IP with DHCP.
This is a very preliminary version which works only once for one
interface.  Nevertheless, it does successfully configure an interface.

Now, why does this need to be inside the kernel, you ask.  To answer
that question, look at all the #ifdefs and platform specific calls that
dhcpcd makes, and especially all the types and macros that are passed
from the user to the kernel and back.  Now, look at all the calls that
are necessary for the application to make to the in-kernel port (hint:
currently rump_pub_netconfig_dhcp_ipv4_oneshot(ifname)).  The former is
seriously unportable.  The latter is 100% portable.