Always set LDSCRIPT=no if KERNONLY is true.
This brings in the assumption that KERNONLY is only used in non-dynamic
environments.  I'm not sure it's a very good assumption, but I think
it's currently true and we need it for things to work.  This should be
properly fixed when we chop up buildrump.sh into smaller pieces and build
it "from scratch" instead of trying to add hacks to make the inherently
posixy platform buildrump.sh support other targets as well.