Attempt #1 at buildrump.sh toolchain flags.
The intent is that the flags from toolchain-conf.mk are included by
entities wishing to use the same toolchain as buildrump.sh to produce
the same kind of binaries.  For example, -m32 or some ARM ISA can be
passed this way without the end-using entity having to "re-know" how to
invoke the tools.

work in progress, not stable 'till the fat binary sings