Change logic for "what is cross toolchain"
Try to detect a cross-compiler based on if $CC contains "-".  Notably,
"cc" might still be a cross-compiler in the strictest sense of the
word if we're e.g. building rumprun.  Here, the driving logic is to
make locating the rest of the toolchain easy without the user having
specify everything.  For example, if CC=cc is given, then it's a good
bet that NM=nm.  Likewise, if CC=foo-bar-cc is given, then it's a good
bet that NM=foo-bar-nm.